refactor: signup page 컴포넌트 분리
메인 섹션 컴포넌트를 추가로 분리해 page 파일 내부 가독성을 높임
